Fix clang compile warnings.

+++ transcoder/basisu_transcoder.cpp
@@ -19973,11 +19973,11 @@ namespace basist
 			if (new_exp > 15)
 				e = 31;
 			else if (new_exp < -14)
-				m = lrintf((1 << 24) * fabsf(fi.f));
+				m = (int)lrintf((1 << 24) * fabsf(fi.f));
 			else
 			{
 				e = new_exp + 15;
-				m = lrintf(flt_m * (1.0f / ((float)(1 << 13))));
+				m = (int)lrintf(flt_m * (1.0f / ((float)(1 << 13))));
 			}
 		}
